Starting Splash page
Got the Splash page working. It was a bitch and I had to kind of hack
it together. I has to take out our mouse click events and put them into
a separate file so if the user clicked on the splash page it would
record any sounds or clicks.

I also cleaned up the file a bit.